Move validateField tests to directory matching source directory name
